deepfake-shield/backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
sqlalchemy
pyjwt
passlib